requirements.txt
requests
gspread
oauth2client
python-dotenv
google-genai
